[Fix] #924 - 리뷰 반영
